+++ chat-service/src/main/java/webrtc/v1/chat/repository/ChatLogRepository.java
@@ -5,5 +5,5 @@ import webrtc.v1.chat.entity.ChatLog;
 import java.util.List;
 
 public interface ChatLogRepository {
-    List<ChatLog> findChatLogsByChannelId(String channelId, Long idx);
+    List<ChatLog> findChatLogsByChannelId(String channelId, Integer idx);
 }
